# Improve script speed with -n option to iptables command