# Description
The security check using the iptables command is slow because it performs DNS resolution, which is unnecessary. The command iptables -L takes around 3 minutes to complete due to these resolutions, which are not required for this check. By adding the -n option, DNS resolution is avoided, improving execution time.

# Proposed changes
Add the `-n` option to the iptables `-L` command to skip DNS resolution and speed up execution.

# Results time
* Without the `-n` option (default):
```bash
sudo ./vps-audit.sh
0.00s user 0.01s system 0% cpu **3:25.20 total**
```

* With the `-n` option:
```bash
sudo ./vps-audit.sh
0.00s user 0.01s system 0% cpu **9.146 total**
```