@vernu 2025-01-17
@alvarosaavedrau got it, thanks for clarifying