@vernu 2025-01-17
> Thank you for reviewing my PR! I understand your surprise about the improvement in execution time; here’s more context to clarify.
> 
> ### Environment
> * **System:** Debian 12 bookworm
> * **iptables rules:** My server has several rules, because I use Docker to run some containers, also I have some manual rules.
> 
> ### Issue Details
> When `iptables -L` is executed without the `-n` option, it attempts to resolve the IP addresses to hostnames. In environments with many rules or slow DNS resolution, this significantly impacts performance. For example:
> 
> * Without `-n`: Takes ~3 minutes (205 seconds) to complete.
> * With `-n`: Takes ~9 seconds, as it skips the DNS resolution.
> 
> The `-n` option is documented as a standard way to speed up commands that list rules, especially in complex environments like mine.
> 
> ### Steps to Reproduce
> 1. Add a large number of rules to `iptables` (e.g., via Docker or manual configurations).
> 2. Execute `iptables -L` and measure the time.
> 3. Compare it to `iptables -L -n`.
> 
> ### Why This Change?
> The `-n` option skips DNS resolution, which is unnecessary for this security check. The primary goal of the check is to verify if rules exist, and resolving hostnames does not add value. Avoiding this step improves performance without altering the script's behavior.
> 
> ### Benefits
> * **Performance:** Users with similar environments will experience faster checks.
> * **No downside:** The `-n` option does not change the output format relevant to the check.
> * **Best practice:** Skipping DNS resolution is common for performance reasons in similar tools.
> 
> I hope this additional context clarifies the rationale. Let me know if I can provide further details or run additional tests. 🚀🚀🚀

got it, thanks for clarifying